* master:
  [gf2x] missing (ugly) piece from last commit
  several changes, incl. some for 32 bits.
  [utils] some docs for params.[ch]
  [duplicates.c] beginning of review: removed AGRESSIVE_MODE=1 stuff (now only
  [utils/relation.c] moved fread_bug function from duplicates.c to relation.c,
  [cadofactor.pl] replaced > $prefix.nodup into -out $prefix.nodup in duplicates
  [purge.c] got rid of warning "right shift count >= width of type",
  reverting purge.c to 1685

+++ gf2x/mul-gf2x.c
@@ -18,11 +18,12 @@
   02111-1307, USA.
 */
 #include <string.h>
+#include <stdint.h>
 
 #include "gf2x.h"
 
 #ifdef MUL_FFT_TABLE
-long T_FFT_TAB[][2] = MUL_FFT_TABLE;
+int64_t T_FFT_TAB[][2] = MUL_FFT_TABLE;
 #endif
 
 /* This is the toplevel multiplication routine. It handles the temporary
@@ -121,7 +122,7 @@ void mul_gf2x_r(ulong * c,
     } else {
 	ulong *v = pool->stk + toomspace(sa);
 
-	long i, j;
+	unsigned int i, j;
 
 	memset(c, 0, sc * sizeof(ulong));
